galaxy/api/serializers/collection.py
from rest_framework import serializers


class UploadCollectionSerializer(serializers.Serializer):

    file = serializers.FileField(
        help_text="The collection file.",
        required=True,
    )

    sha256 = serializers.CharField(
        required=False,
        default=None,
    )
