requirements/dev-requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements/dev-requirements.txt requirements/dev-requirements.in
#
aiohttp==3.5.4
alabaster==0.7.12         # via sphinx
amqp==1.4.9               # via kombu
ansible-lint==4.0.0
ansible==2.7.4
anyjson==0.3.3            # via kombu
asn1crypto==0.24.0        # via cryptography
async-timeout==3.0.1      # via aiohttp
atomicwrites==1.2.1       # via pytest
attrs==18.2.0             # via aiohttp, pytest
babel==2.6.0              # via sphinx
bcrypt==3.1.4             # via paramiko
billiard==3.3.0.23        # via celery
bleach==3.0.2
celery==3.1.24
certifi==2018.11.29       # via requests
cffi==1.11.5              # via bcrypt, cryptography, pynacl
chardet==3.0.4            # via aiohttp, requests
click==7.0                # via dynaconf, rq
coreapi==2.3.3            # via drf-yasg, pulpcore
coreschema==0.0.4         # via coreapi, drf-yasg
coverage==4.5.2           # via pytest-cov
cryptography==2.4.2       # via ansible, paramiko, pyopenssl
defusedxml==0.5.0         # via python3-openid
deprecated==1.2.4         # via pygithub
dj-database-url==0.5.0
django-allauth==0.34.0
django-bootstrap-form==3.4
django-celery==3.2.2
django-debug-toolbar==1.11
django-filter==2.0.0
django-log-request-id==1.3.2
django-prometheus==1.0.15
django==2.1.5
djangorestframework-queryfields==1.0.0  # via pulpcore
djangorestframework==3.7.7
docutils==0.14            # via rst2html5-tools, sphinx
drf-nested-routers==0.91  # via pulpcore
drf-yasg==1.12.1          # via pulpcore
dynaconf==1.2.0           # via pulpcore
flake8==3.6.0
gunicorn==19.7.1
honcho==1.0.1
idna-ssl==1.1.0           # via aiohttp
idna==2.7                 # via cryptography, idna-ssl, requests, yarl
imagesize==1.1.0          # via sphinx
inflection==0.3.1         # via drf-yasg
influxdb==5.2.1
itypes==1.1.0             # via coreapi
jinja2==2.10              # via ansible, coreschema, sphinx
jog==0.1.1
kombu==3.0.37             # via celery
markdown==2.4.1
markupsafe==1.1.0         # via jinja2
marshmallow==2.16.3
mccabe==0.6.1             # via flake8
mock==2.0.0
more-itertools==4.3.0     # via pytest
multidict==4.5.2          # via aiohttp, yarl
oauthlib==2.1.0           # via requests-oauthlib
packaging==18.0           # via sphinx
paramiko==2.4.2           # via ansible
pathspec==0.5.9           # via yamllint
pbr==5.1.1                # via mock
pluggy==0.8.0             # via pytest
prometheus-client==0.5.0
psycopg2-binary==2.7.6.1
pulpcore==3.0.0b18
py==1.7.0                 # via pytest
pyasn1==0.4.4             # via paramiko
pycodestyle==2.4.0        # via flake8
pycparser==2.19           # via cffi
pyflakes==2.0.0           # via flake8
pygithub==1.43.3
pygments==2.3.0           # via sphinx
pyjwt==1.7.1              # via pygithub
pynacl==1.3.0             # via paramiko
pyopenssl==18.0.0
pyparsing==2.3.0          # via packaging
pytest-cov==2.5.1
pytest-django==3.4.4
pytest-env==0.6.2
pytest==4.0.1
python-box==3.2.3         # via dynaconf
python-dateutil==2.7.5
python-dotenv==0.10.1     # via dynaconf
python3-openid==3.1.0     # via django-allauth
pytz==2018.7
pyyaml==3.13
redis==3.0.1              # via rq
requests-oauthlib==1.0.0
requests==2.20.1
rq==0.13.0                # via pulpcore
rst2html5-tools==0.5.3
ruamel.yaml==0.15.85      # via drf-yasg
semantic-version==2.6.0
semver==2.8.1
six==1.12.0
snowballstemmer==1.2.1    # via sphinx
sphinx==1.7.9
sphinxcontrib-websupport==1.1.0  # via sphinx
sqlparse==0.2.4           # via django-debug-toolbar
toml==0.10.0              # via dynaconf
typing-extensions==3.7.2  # via aiohttp
uritemplate==3.0.0        # via coreapi, drf-yasg
urllib3==1.24.1           # via requests
webencodings==0.5.1       # via bleach
wrapt==1.10.11            # via deprecated
yamllint==1.13.0
yarl==1.3.0               # via aiohttp
